add : fun write neo4j

+++ src/main/kotlin/io/ioNeo4j/EdgeNeo4j.kt
@@ -14,5 +14,5 @@ class EdgeNeo4j(
     var label: String = "",
 
     @Property
-    var weight: Double? = null
+    var weight: Float? = null
 )
